refactor(langy): mount the health probe beside its siblings under /api/health (#7944)
The Langy canary moves from GET /api/langy/health to GET /api/health/langy,
declared public and authenticating in-handler like the other subsystem
probes, and answering refusals in their { message } shape. The auth chain it
shares with the turn routes is extracted to
app-layer/langy/langyApiKeyAuthorization.ts so both call one function; the
chain itself (credential, surface flag, langy:create ceiling, cohort, actor)
is unchanged.
