+++ platform/app/.env.example
@@ -606,9 +606,10 @@ LANGY_INTERNAL_SECRET=
 # getting a budget of its own.
 
 # LANGY CANARY HEALTH PROBE
-# GET /api/langy/health sends Langy one real greeting turn and reports whether
-# it answered. It needs NO env var of its own: it is authenticated by the same
-# chain as POST /api/langy/conversations (a project API key in X-Auth-Token,
+# GET /api/health/langy sends Langy one real greeting turn and reports whether
+# it answered. It needs NO env var of its own: it sits beside the other
+# /api/health probes and is authenticated by the same chain as
+# POST /api/langy/conversations (a project API key in X-Auth-Token,
 # the release_langy_api_key_turns_enabled flag, the langy:create ceiling, and
 # a key owner in the Langy cohort), and the turn runs as that owner. Failure
 # reasons are exactly three: timeout | turn_failed | empty_reply. "timeout"
